The virtual Apostrophe page associated with a blog post or event has gone through another slug change. Run the apostrophe-blog:migrate-page-slugs task to update your page slugs appropriately. If you are overriding the templates or making your own, start getting the virtual page slug via $a_blog_post->Page->slug rather than by making an assumption about it as we formerly did.
The blog and event admin modules now check for blog_author or admin permissions as a minimum standard for entry. We need to also update the "new post" button and related helper buttons on the public blog and calendar engine pages to check for this. isAuthenticated() is not good enough.



git-svn-id: http://svn.apostrophenow.org/plugins/apostropheBlogPlugin/trunk@1289 c89fab08-2f9b-42bf-8176-bd3ad6bee547