Being a user for a category means that you can categorize your posts into it. It does not mean that you can edit other people's posts that happen to be in it. Only blog admins have blanket permission to do the latter.
You must be a member of the editors group (if there is one) to be listed as a category author. This limits the size of the dropdown and is necessary for the same reasons as in the page settings dialog.

Added an overridable getUseFields method to aBlogCategoryForm::setup() making it possible to add fields without getting clobbered by the useFields() call in setup().



git-svn-id: http://svn.apostrophenow.org/plugins/apostropheBlogPlugin/branches/1.4@1734 c89fab08-2f9b-42bf-8176-bd3ad6bee547