First pass at a calendar for events.
Cannibalized a portion of the calendar from Trinity (EventCalendar.class.php)

The arrows move the calendar from month to month.

Days with events are clickable and take you to that day view.

Removed the browse by day month year interface because it was redundant and felt less intuitive than the calendar now that it is in place.

Dan also fixed a push pop url bug relating to blog and events search results. This related to F&M's bug today.

I need to take another run at the calendar, adding some missing functionality tomorrow.

When I feel good about it, I'll port it over to the trunk.



git-svn-id: http://svn.apostrophenow.org/plugins/apostropheBlogPlugin/branches/1.4@2031 c89fab08-2f9b-42bf-8176-bd3ad6bee547