Added some more functionality to the calendar so you can click on a day, click on a month or year and begin browsing that way.
The left and right arrows on the calendar work.

If you hover a day with events, a tooltip tells you how many events there are on that day.

The current day gets a stroke

The selected day is highlighted with a background color.

All of the styles are light and minimal and the calendar should be lightweight and adaptable to any design pretty easily.

The calendar object itself contains a full event object so the calendar itself could be extended to display more event information if needed.

and lastly, it's backwards compatible with all 1.4 projects. By default the calendar is turned off. To turn on the calendar in app.yml: app_aEvents_display_calendar = true



git-svn-id: http://svn.apostrophenow.org/plugins/apostropheBlogPlugin/branches/1.4@2034 c89fab08-2f9b-42bf-8176-bd3ad6bee547