Implemented #662, hit #476 and #664.
When you click "New Post" or "New Event" you are prompted in-context for a title. If you cancel nothing happens. If you save, a real blog post or event with a real title and real slug comes into being, and you are redirected to the edit page for it. I also refactored a lot of stuff related to how titles and slugs are set. This resolves quite a few chicken and egg problems and mysterious behaviors. Blog and events feel a lot more usable



git-svn-id: http://svn.apostrophenow.org/plugins/apostropheBlogPlugin/trunk@2457 c89fab08-2f9b-42bf-8176-bd3ad6bee547