The events slot is now based on the blog posts slot. I did as much reuse as possible so there's not a lot of code specific to it.
It works, however the styles currently are not finding it, so it's not really usable yet. The typeahead ajax queries for titles are working but you can't see the results, presumably this is styling / js scoping related. The JS to hide the "other" half of the form is not working. I'm assigning that stuff to John as I haven't managed to find it so far



git-svn-id: http://svn.apostrophenow.org/plugins/apostropheBlogPlugin/branches/1.5@2926 c89fab08-2f9b-42bf-8176-bd3ad6bee547