This one is epic.
Fixes #711: you can now combine browsing by date, tag, category and plaintext search. The currently active filters are clearly called out at the top of the main content area where you can't miss them, in the form of a sentence, per discussion last week. It's easy to remove any active filter. 

The list of tags is reduced according to the filters active for everything else (date, category, search). The same has been implemented for categories. Thus even though you are allowed to combine filters you never have a "WTF" moment in which you can click on a tag but nothing comes up. The filters you see are always relevant in the context created by the other filters.

The blog and events actions have been extensively rewritten to use new direct SQL queries that permit all of this with reasonable efficiency. Those live in aBlogToolkit. The direct SQL stuff is implemented via the aMysql class, which I contributed back to Apostrophe from the plog project, which borrowed an early form of it from the Apostrophe aMigrate class (I am opening a ticket to, ahem, migrate all aMigrate calls to aMysql calls).

There is more code sharing between blog and events.

Note that some small tweaks to overridden blog templates will be needed, the sidebar partial takes slightly different arguments and there is a new filters partial to be called.

PLEASE NOTE: there should be "back" and "forward" arrows for browsing blog posts and events by date, and they are there but are not visible. John is aware of this CSS issue and has a separate ticket to fix it. I mention this because it will probably cause some confusion when evaluating the new "filters list as a sentence" feature.



git-svn-id: http://svn.apostrophenow.org/plugins/apostropheBlogPlugin/branches/1.5@2952 c89fab08-2f9b-42bf-8176-bd3ad6bee547