Rocked #1064: blog posts can now be filtered by author.
Also fixed several cases in aBlogToolkit::filterForEngine where the inappropriate use of an INNER JOIN was precluding the display of alternate choices for some filters in some situations. Right now we always address that by using left joins and adding WHERE clauses to the individual queries for the other filters at the end of that method. There is probably a performance win to be had later by coming up with a maintainable way to do INNER JOINs for some queries and LEFT JOINs for others but I don't want to do that piecemeal as the method is already significantly complex and the 1.5 blog plugin's filtering performance is already a big improvement on 1.4



git-svn-id: http://svn.apostrophenow.org/plugins/apostropheBlogPlugin/branches/1.5@3961 c89fab08-2f9b-42bf-8176-bd3ad6bee547